requirements.txt
Pillow>=8.0.0
requests>=2.25.0
tkinterdnd2>=0.3.0